rozwiązanie2 - zrobiłam poprawki ale niestety nie mogę wysłać- hasło w properties ustawiam poprawnie
