+++ src/main/java/com/example/zadanie221/MailService.java
@@ -1,6 +1,8 @@
 package com.example.zadanie221;
 
+import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.slf4j.event.EventRecodingLogger;
 import org.springframework.mail.MailSender;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -9,11 +11,11 @@ import org.springframework.stereotype.Service;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
-import java.util.logging.Logger;
 
 @Service
 public class MailService {
     private JavaMailSender javaMailSender;
+    private static final String SENDER = "forsale1@op.pl";
 
     public MailService(JavaMailSender javaMailSender) {
         this.javaMailSender = javaMailSender;
@@ -24,7 +26,8 @@ public class MailService {
         try{
             MimeMessage mimeMessage = javaMailSender.createMimeMessage();
             MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, "utf-8");
-            helper.setFrom(dto.getReceiver());
+            helper.setTo(dto.getReceiver());
+            helper.setFrom(SENDER);
             helper.setSubject(dto.getTitle());
             helper.setText(dto.getContent(), true);
