+++ src/main/resources/templates/error.html
@@ -1,15 +1,14 @@
-<!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8">
-    <title>Potwierdzenie</title>
+  <meta charset="UTF-8">
+  <title>Potwierdzenie</title>
 </head>
 <body>
 <div th:replace="fragments :: header"></div>
 <h3>Potwierdzenie wysłania wiadomości</h3>
-<p>Dziękujemy za kontakt. Odezwiemy się w najbliższym czasie.</p>
-
-
+<p>NIE UDAŁO SIĘ WYSŁAĆ WIADOMOŚCI.</p>
+<p>Spróbuj jeszcze raz:</p>
+<a href="/kontakt">Kontakt</a><br>
 <br>
 <a href="/">Powrót do strony głównej</a><br>
 <div th:insert="fragments :: footer"></div>
